Fixed issue 40196: [RR19Q1] [OMNI] Usability of the new "Pay Open Tickets" pop up should be improved
Now, before do anything, we check all the orders that the client check in the popup.
An array is created with all the paid tickets that it has been found.

if all the selected orders are paid, An specific popup is shown just saying that at least
one of those orders must be a not-completly-paid order.

if anyone can be paid, we follow the pay open tickets flow with all of the ticket that we
can pay