fixes issue 40255: The woodstox library is not required at runtime
The woodstox library which provides an implementation of the Stax XML API is used by DBSM for export/import taks but it is not required at runtime. Therefore, now we are not copying it when deploying the context.

--HG--
extra : rebase_source : d5a8804bdefd828af6fc8cf3fe1460810e7828fd