Fixes issue 40275: Error when calculating the field "amount" discount
The discount was wrongly calculated because an incorrect rounding on its calculation.
When discounts are applied the ActualPrice is calculated from the standard price and rounded.
When the reverse process is made (calculate the Price Std from the Actual Price) could be
differences because the rounding previously done to the ActualPrice.
To avoid those differences first is divided by the price list and rounded and then multiplied
by 100 and rounded again when the discount is calculated.

--HG--
extra : rebase_source : e82343f8c97581af972008e85d821196c1cbbe58