Fixed issue 38791: Using comma as group separator causes problems in CSV export.
Numbers in csv does not need to have formatting, and using group separators
makes no sense. For this reason, number formatting is ignored when exporting to
csv and only the comma separator is taken into consideration to determine
whether to use a comma or a semicolon as the field separator.

--HG--
extra : rebase_source : eee18115838de741a5b2022a3f26a6c401be5435