Fixed issue 40299 : ExternalOrderLoader marks order lines as delivered during the Create step
The property obposQtytodeliver must be set in the external order loader in order to set correctly the delivered quantities.
In case that the order is being created or paid, the quantity is taken from the delivered quantity or zero if there's not delivered quantity.
In case that the order is being delivered, the quantity set is the ordered quantity.
If the object was already sending the property, the value is maintained, allowing to have partial deliveries.
Added also the correct properties to be able to invoice and deliver when the order is synchronized due to the step ALL.