Fixed issue 40382: The paid amount of the invoices is not set correctly in some cases
When checking each payment detail if is counted for the amount to pay, previously the check was being done using a static constant instead of a value in each payment. Now that check is done comparing the payment status with the status in the financial account.

--HG--
extra : rebase_source : b0f4b9f704e1477c73404e5f6d036d3456d99b35