Fixes issue 40392: Can't modify an invoice payment plan if it's already modified
This solution takes into account all the Payment Plan lines when calculating that
modified outstanding amounts are modified.
Also, all the payment plan expected amounts are updated if the sum of the updated
outstandings amounts are equals than original outstanding amount of the invoice instead
of just update the last record.

Also when a new record is added are taken into account all the lines and the new
outstanding amount is calculated by subtract to the total expected the total already paid
and the total outstanding in the existing lines.

--HG--
extra : rebase_source : 2d4bd6456a4cc20e842356941399c36cbd1644b5