Fixes issue 40186: Pending Goods receipts with movement quantity=0
The fieldTotalQty and fieldQtyordered fields were incorrect defined to use "euroEdition"
instead of "qtyEdition".
Columns for Ordered Quantity and Movement Quantity of the lines of Purchase Order and Goods Receipt
windows are defined with a "Quantity" reference and it is implemented internally to
format the values using the "qtyEdition" setting.
To do more homogeneous the way the quantities are shown in the process and in the different
windows, all of them will be formatted by the same way: using settings of qtyEdition property.