Fixes issue 40333: Advance Filters in Trial Balance is not working
The account from/to filters should be shown just if the Account Level filter is
"Subaccount" in any other case it should be hidden as they are used in the report
just for this particular case.

The display logic of this section/fields was incorrect defined because after a previous
search it was setting as TRUE to display the fields if the Advanced Filters section is
expanded, but it might take into account also the display logic explained before
asking if the selected account level is "Subaccount".

When the Advanced filters section is expanded then these fields will be hidden if the
selected Account level is different than "Subaccount"