Fixes issue 39167: Characteristics labels hyperlink in Update Characteristics popup can be clicked and raise a server error
Due to an incorrect definition of the target entity and reference column the navigation
from a Characteristic value in the Update Characteristic popup it was unable to redirect
to the right window and an exception was thrown in the backend.
This solution define 'CharacteristicValue' as the target entity and 'name' as referenced column.

--HG--
extra : rebase_source : 4432d0f12da34120f463db124d2cb99f1cf26b10