Fixed issue 40601: The calculateReceipt function is called to much times when changing prices
New API to set prices to lines. It should be always called, even if the call is only for a single line.
Removed the call of the calculate receipt done after the price is modified. Now, when updating a single line's price it is only called from the event launched when updating the price of the order line.
For multi selection, the calculateReceipt is disabled until the last price is updated. After that, it is enabled and launched.
Same modifications for the undo action.

--HG--
extra : rebase_source : dde0d2448c9ec1a00052b9523f905ef7bdfe2c61