Fixes issue 40519: Prevents post CostAdjustment with closed periods
For CostAdjustments, the accounting date is a property of CostAdjustmentLines,
the date considered to get the period is the reference date of the document,
causing a CostAdjustment with reference date in an open period can be posted
even when there are lines with accounting date in a closed period.

This fix prevents a CostAdjustment to be posted if there is at least one line
with accounting date in a closed period.

--HG--
extra : rebase_source : 790949bacc62778ade7d0a5f06da8e2ced7e89b0