fixes issue 40743: wrong export to CSV having the grid filtered by a date time
After the following changeset[1], exporting the grid to CSV was not working properly having the grid filtered by a date time column. Although the fix is correct, it caused the criteria sent to the datasource in this type of requests to stop being handled properly.

To avoid this problem, when exporting to CSV a grid filtered by a date time column we send in the criteria the plain date (in the same way as it is being done for the standard grid DS requests) and the datasource will take care of building the date filter properly through the AdvancedQueryBuilder.

[1] https://code.openbravo.com/erp/devel/pi/rev/90ee5c965823996a7b4c0ae17d3a3ca71e3fc703

--HG--
extra : amend_source : 1c9eec52b7f163fae1fdc5e941461e94e9ef197f