Fixed issue 40791: Roles require back-office access to log in to mobile apps
MobileLoginHandler does a back-office login to later select mobile default role,
it expects back-office to fail and keep session in case of restricted role login
attempt.

Back-office login handler was changed to fail and remove session in case there is
no back-office role in current user's list. This made not possible to log in to
mobile apps.

Now if logging in to a mobile application session is kept after failure as it was
before.