Fixes issue 40763: Standard properties are taken from underlying window in P&E
When a P&E window contains a grid and a request to the datasource of the P&E window is done, it is
not clear whether the P&E grid or the underlying grid standard properties (tabId, tableId, windowId, etc)
should be included in the contextInfo.

Before this changeset [1] was pushed, the standard properties were taken from the underlying grid. The
changeset modified the behaviour, so that the P&E grid standard properties were included instead.
This was necessary because the SelectorConstants.includeOrgFilter method needed the table ID of the
P&E grid.

But the change caused another problem: Some OBBindings were relying on the underlying grid standard
properties being included in the contextInfo.

To fix both problems, now a new property called inpPickAndExecuteTableId is included when doing a
request from a P&E grid, and that property is referenced by SelectorConstants.includeOrgFilter.

The contextInfo now will include the standard properties of the underlying grid, as it was done
before [1].

[1] https://code.openbravo.com/erp/devel/pi/rev/1ed089db27e7036cf53026ca873dbcf104d79bb9

--HG--
extra : rebase_source : 5ceaff15c735bc050792c1f828fce92b7cb474a3