Fixes issue 39865: Do not create line if mandatory fields of current are empty
In the view grid, if in a new line there are mandatory fields not filled in, it should not be
possible to create a new line.

To prevent that, the updateState function of the new line toolbar button has been updated, and
the doCellEditEnd function of ob-view-grid also takes it into account.

--HG--
extra : rebase_source : 1ac425d171c3d9f9ee3df661515551f3951a065a