Fixes issue 40264: Ignore some invalid functions when loading the database model
Openbravo does not support the definition of functions with anonymous parameters. Before this
changeset, loading from the model one of those functions resulted in a NullPointerException.

Now these functions will be ignored when loading the model from the database. This means that
export.database will not export them, and update.database will not overwrite them.

To be able to ignore a row within RowConstructor.getRow a new exception called SkipRowException has
been defined.