Fixed issue 33973: implement cashvat in webpos
Fill the fields paidAmtAtInvoicing (c_invoice) and isPaidAtInvoicing (fin_paymentdetail),
used in the backend to calculate the cashvat, and to which gl item correspond the payment

When having prepaiments (order with payments and after created the invoice) call to
CashVATUtil.createInvoiceTaxCashVAT api to fill the cahsvat table. This is done when the invoice
is created in the pos or in the cashup.

Needed to add to createPSInvoice missing updates of hibernate objects, so the later call
to createCashVat works.

When the organization is set as cashvat send it to the webpos in the terminal info.
Added also to the ticket json a field cashvat when the organization/terminal works with cashvat

When returning a ticket it should mantain the cashvat value of the original ticket, if
Example: if an organization change to cashvat, when returning a ticket from before the
change it should return as not cashvat

The taxes available when cashvat should be only cashvat ones and exemption ones.
If not cashvat, it should not be available any cashvat tax.

--HG--
extra : rebase_source : 0e604a7f6e72cb6e2a6fefd005142d9d84db4a04
extra : amend_source : a263d0b0f9a6d1a3da2a3c00f8de610fd3c0fcef