Fixes issue 40970: Manager Variants has now proper prod characteristics filter
The ProductCharacteristicsDS was not able to build a proper where clause when the product characteristics
filter was part of a grid in a process definition window.

A new hook called ProductCharacteristicCustomWhereClause has been added, where it is possible to define
custom where clauses for each Process Definition that contains a grid with a product characteristics
filter.

For that hook to work, the request to ProductCharacteristicCustomWhereClause now includes the process id
and the context info of the tab that owns the process definition button, if applicable.

--HG--
extra : rebase_source : 95e9c08d121aae110ab25a1ed7832235c5afd9d8