Fixes issue 40970: Uses ApplicationScoped properly, makes code more clear
The @ApplicationScoped annotation is not inherited, so it must be included in the subclasses, not
in the interface definition.

Also, replaces:

return this.Super('transformRequest', arguments);

With:

return this.Super('transformRequest', dsRequest);

To make the code more clear.

--HG--
extra : rebase_source : a42fec0d0c07d93c7f2956e7656772154c00fc37