Fixed issue 41047: The deferred service can generate the invoice now
When a deferred service is delivered due to an order loader hook, the invoice is not created in the case that should. Now, it will be possible to load again that order in the Web POS and synchronize it again creating the invoice for the deferred service lines.

--HG--
extra : amend_source : 9b5da2b77ddea30292842b18507271c8b8004145