Fixed issue 41196: When the multi order process fails in the preOrderSave hook, it is not correctly managed
Now, when the pay open tickets flow fails during the preOrderSave hook, the order data is restored depending on the synchronized mode status.
If the synchronized mode is disabled, instead of restoring the order model to the frozen model, only those needed properties are set to the initial status.
In synchronized mode, the full order is restored. This means that the backbone model is fully changed. In addition, now the order list model is also restored (updated) to the frozed model, maintaining the same object in both orderList and multiOrder lists.
In addition, payments are always fully restored to the frozen status. Up to now, only the 'origAmount' was restored, having, for example, the property 'amount' with value zero after the values were restored.

If the preOrderSave hook fails, the Web POS waits until all orders have been treated until they are restored, to avoid to restore data in and then to modify the order after that.