Fixed issue 41271: Layaways can generate shipments when the synchronization process have some failure
Now, the 'completeTicket' property is set once the drawer has been closed, to avoid the possibility to set, close the drawer and then modify the ticket before synchronizing. Now the property will be set when the ticket is going to be synchronized.
In addition, if the flow is now stopped, the property is unset, and will be set again if the ticket is synchronized again in the correct conditions.

--HG--
extra : rebase_source : d6540fe9a393b7fb8ebe76bffd7abbeb055aba26