Fixes issue 41352: Runs validation after modifying value in non-editing row
In pick and execute grids, when a row other than the one that has the editGrid is programatically edited
( for instance using the setEditValue function), the validation should be run again. Otherwise the row would
not be revalidated, since validation is done when leaving the row that has the editGrid.

--HG--
extra : rebase_source : e2047aba014a801b7064ba8a4d9ad39ca943c076