Fixes issue 41352: Validates row when it is modified programatically
The problem was that when a row was programatically modified using the setEditValue method, the
change was not validated unless it was done in the row that was currently being edited in the UI.

Now the change will be validated, but only if the row is selected.

A changes has also been done on ob-aprm-addPayment.js to deselected a row before updating it,
because otherwise it caused a validation error.

--HG--
extra : rebase_source : d51fbf685bda0f9469c07f1caab1b50c42a27280