Fixed issue 41407: Changes to make the stock queries responses more extensible
Now it will be easier to manage the stock responses from external modules, without affecting to the possible response other stock check.
In addition, code performed to avoid duplicity.
Update the warehouseqty value received from the StoreDetailedStock.java class, to have the value as a numeric value, not as a string value.
The 'payOpenTicketsValidation' now instead of having a 'loading' scrim, it is used with an 'opaque' scrim. This allows to show different popup during the process to set the needed validations. It was not possible to do it with the 'loading' scrim.