Fixes issue 41215:Error using FIN_Utility.getDocumentNo in an event handler
An error occurs when FIN_Utility.getDocumentNo is called for a sequence that is
referenced in the entity being saved. When this is done, an NPE will happen in hibernate.

The problem is related to the call of lockSequence(seq) in getDocumentNo.
The specific problem is OBDal.getInstance().getSession().evict(seq)

The evict call was removed and also some code refactor was done

--HG--
extra : rebase_source : e30a0e1627c199433eec9fcb1810bcc76b354128