Fixes issue 41664: Record added in filtered empty grid is not filtered out
This code [1] was already in charge of manually adding to the grid records that have been created
in the form but that do not comply with the grid filter.

But that code was not working properly if the grid was empty when just before adding the new
 record, because in that case view.viewGrid.body.getVisibleRows() (which is used to determine where
the new row should be placed) returns [-1, -1] (meaning that at that moment there are no visible
 records), so the visibleRows[0] !== -1 was not met.

An extra else clause has been added to handle the case where the grid total rows is 0. In that case
the new record is added in the position 0.

[1] https://code.openbravo.com/erp/devel/pi/file/a3916b3d436d/modules/org.openbravo.client.application/web/org.openbravo.client.application/js/form/ob-view-form.js#l2013

--HG--
extra : rebase_source : 9ca2ef7bee974f016cbfaf46d88c0825a24828f3