Fixes issue 41694: Wrong accounting entry for Cost Adjustments over production products with transactions in more than one legal organization
When a Cost adjustment have lines from with organizations that doesn't belongs to
the same natural tree a wrong accounting schema was used to create the fact lines.

Now the accounting schema used to create the fact lines are checked if it's belongs
to the same natural tree than the transaction line.