fixes issue 40746: When subtab process deletes parent record view is refreshed
When a subtab process deletes its parent record, parent view is refreshed correctly.
If parent was in Form view, it is switched to grid view, an info message is shown and the grid refreshes to show the change.
If parent was in Grid view, view is refreshed to show the changes.

--HG--
extra : rebase_source : 548ad8c21f3cc16b65feda13d32086816c5439a6