Fixes issue 41909: Fixed BP filtering on Create sales order from expenses
If the Business Partner had an inactive contact, query loading records of the BP
selector was not taking into account the business partner itself. It was caused
by an incorrect filtering at WHERE CLAUSE for those business partners without a
contact or with an active contact, instead of doing at join time with the AD_USER
table.

The solution was to move the condition (u.isactive is null OR u.isactive='Y') to be
validated at the JOIN with the AD_User table, so if there is a contact but it is inactive
then information of the contact will be not retrieved, but that related with the business
partner yes.