Fixes issue 41935: Take into account total paid when calculating the outstanding amount of lines when using the Modify Payment Plan feature.
In some cases, using the Modify Payment Plan feature was shoing that the Outstanding Amount
was invalid because the total invoice amount was considered when validating the Outstanding Amount
but wasn't taking into account the other payment plans already paid.
The issue on this situation was to to get the received amount from records shown in
the grid, as those lines already paid are not shown in the pop up grid, the amount
don'tr match and an error is shown to the user.

To fix this issue is validated that the total Outstanding amount is exactly the difference
between the invoice grand total and the total invoice amount already paid