fixes issue 40836: Generates deprecated tag and Javadoc for DAL entities
Generates deprecated tag and Javadoc @deprecated message for DAL entities that have set "Development Status" variable to Deprecated.
Also sets as deprecated getters and setters in entities that target deprecated entities.
OB property hb.generate.deprecated.properties must be set to true for the generation to happen.
If table is set to deprecated, generated entity class is marked as deprecated.
If column is set to deprecated, generated entity method corresponding to that property column is marked as deprecated.