Fixes issue 41954: Edited selected records refreshed after gid is reloaded in Add Payment process
When the Transaction Type option is changed the Order/Invoices grid is reloaded
and the records are removed and loaded with possible changes according selected
option. To fire all the amounts updation logic is fire the doSelection event on
edited selected records of the grid, and this action will update the amounts with
the correct values.

--HG--
extra : rebase_source : 53afa2d6955ce4fbc24387b254a46ecc33187938