Fixes issue 42057: It is not possible to unpost an Internal Comsumption
If the Internal Consumption documents have not Document Type defined which is mandatory
when unposting a document an exception will be thrown to the user.

Also was included a minor change in the Posted class to set the success message title
after we are sure any exception was thrown to avoid to have an error message with a success
title and cause confusions.

--HG--
extra : rebase_source : b0643797f9eb7d99dfeeb69ff635832d2fc0c12b