Fixes issue 42051: Removed all the keepOnAutomaticRefresh flag activation as it is not necessary and it is changing the default behavior of any message shown in the application.
The keepOnAutomaticRefresh was set to true in all the cases inside the cancelAndReplace
EventHandlers, no matter if a message was shown or not, and this causes that next message shown,
even if it is not related to the Cancel and Replace, it is fixed and is not automatically removed.

The keepOnAutomaticRefresh flag is used to avoid a message to be automatically removed but it is
not needed on this functionality and it was removed.