Fixes issue 42092: Balanced posting when adjusting a return receipt
When a Goods Receipt Return was adjusted a balance in Warehouse Differences and
Invoice Price Variance account was generated. As only the cost adjustments generated
by Price Difference Correction process are posted on the Invoice Price Difference account
all other adjustments done to receipts are posted on Warehouse Differences account.

The correct behaviour should be: If the adjustment is unit cost the account used should
be the Invoice Price Difference account independently of the cost adjustment source process.