Fixes issue 41880: Actual Payment updated in Match Statements for operations with different currencies.
Fixed the updation of the actual payment with the value of the bslamount converted
to the same currency and using the defined conversion rate.

Created a new function to update an amount with the current currency and precision.
The new converted amount shold be calculated multiplying or dividing by the currency
exchange rate according the conversion is direct or inverse.

--HG--
extra : rebase_source : 3b6c8efac90f78e4dfef7e18749c5fd5469ddbd5