Related to Issue 42430. Changed validation in Trigger
The validation in the trigger did not allow to do any change to an
already processed or posted inoutline.

The change done is to allow to modify some columns and to control
which ones can not be changed. For example, the updated column
can be modified after this change even if the inoutline is already
posted or processed.

--HG--
extra : rebase_source : bdd8ac844681c576c89c47c814243b8b9b5c07dd