Fixed issue 42116: Verify prepayment limit and pick&carry lines at the same time
When partially paying a ticket, the validations for the prepayment percentage limit
and delivery modes should be done simultaneously showing always the lowest amount
needed for a successful payment. So when the check for the partial payments' preference
fails, we need to still check for the Delivery Modes preference and inform the user
whether the payment can be accepted or not, and the lowest allowed quantity.

--HG--
extra : rebase_source : 6501a128f2ee78e6bea872b968e652953c11c65d