[PaymentRounding] Payment Rounding line can be removed after POS reload
When a payment rounding is created and the POS was reloaded the payment rounding
line can be removed.

Now when the POS is reload the payment rounding line is related with the new payment
created to avoid that can be removed.

--HG--
extra : amend_source : 9687bbb6c58f925c00643ac0d48bea02835579a6