Fixes BUG-42520:Pareto Product Report shows wrong stock
** Stock quantity was added as many times the no of records
in m_transaction_cost related to m_transaction, to fix
issue m_transaction_cost is removed as main table in select
query, m_transaction is used as main table in the select
query, a subquery is used to calculate sum of the cost from
m_transaction_cost taking into consideration currency conv
if required, then this subquery is join with m_transaction
main table in the query. While re-arranging the queries,
maxaggDate parameter order is changed as required.