Fixes BUG-42595: Don't cache StyleSheets if any module in development
Previously StyleSheets were cached if working on a module and there weren't any other module with CSS set to inDevelopment

Behaviour for checking if instance is in development is now the same for both StyleSheetResourceComponent and StaticResourceComponent. It follows BaseComponent isInDevelopment logic, if a module is in development, all the instance is in development.