Fixes BUG-42915: Tree selectors can be used in menu parameter windows
The problem was that the code in OBTreeItemPopupFilterWindow.copyFunctionsFromGrid was assuming that the
grid where the selector has placed had a view that also had a viewGrid. This is not the case in grids that
are parameters in parameter windows that are open directly from the menu.

What the function does is to copy from functions from OBViewGrid to OBTreeItemPopupFilterWindow. For this
it does not need any real instance of viewGrid, any instance works. Because there is no real instance
available in this type of parameter windows, the issue has been fixed by creating a dummy one like this:

        if (grid == null) {
          grid = isc.OBViewGrid.create({ fields: [], view: {} });
        }