Fixed issue 42911: Correctly applied promotions to product with characteristics
Promotions or discounts for a concrete characteristic was not being applied well
because the first time the product was added to the ticket, a new line was created
and finally as the last step, the product characteristics was updated to the line.
The discount engine was verifying if was possible to apply the discount for products
with characteristics defined in the rule, but it was happening before the product
characteristic list of the product was updated in model, so the rule was been marked
as not possible to apply to the new created line.

To fix this issue, was updated the product model to be added into the ticket, with
all it characteristics before adding the product to the line, and execute the discount
validations.
This fix updates the productCharacteristics attribute of the product model before
the line is created, instead of update the line itself after the line was created
and discounts validations fired.

As the product characteristics are updated previously to create the line (before
the OBPOS_PreAddProductToOrder hook is executed), it is not needed to do this when
the OBPOS_PostAddProductToOrder hook is executed, so the execution at this point of
the code was removed.