fixes BUG-43201: Inconsistent grid state after editing a child record
  After editing a child record of a tab that has an HQL where clause
with dynamic variables, the request to refresh the parent record was not
working fine.

  This is because the HQL where clause could not be resolved because the
context information is not sent in this kind of request.

  But in this case we do not need to include the tab HQL where clause at
all, as we want to refresh just the record and that means that we are
going to filter by the record ID.

  Therefore, to fix this issue we are setting the _directNavigation and
_targetRecordId flags to avoid including additional filter clauses in
the request to the DefaultJSONDataService. Thus, in the request built to
refresh the current record we are now filtering just by the record ID.