[SafeBox] Implement logic from cash up process with Safe Boxes
Cash up for payment method of Safe Boxes works different than other
payment methods.

During the Cash Up process in WebPOS, payment methods set as “Defined in
Safe Box” are not counting during the process.

When the Cash Up process ends in WebPOS and is synchronized. the full
amount of transactions (deposits, drops, returns) made by a Safe Box
payment method is moved to the financial account defined in the Safe Box
payment method. Doing so, the ending balance for the reconciliation of
that payment method is 0.