[SafeBox] Refactor CashUp components and process flow
The main idea of the Refactor is to separate and be able to reuse
the maximum amount of code and components for Safe Box counting
process.

To be able to distinguish the especific cashup and countsafebox
flows, several new folders have been added inside close cash folder.

The common logic for both process has been kept in the root location
of closecash subfolders.

Several components has changed the name (for the common components
). The CSS classes has been changed also to follow CSS naming rules
of OB.

The affected logic now uses ES9.