Added the possibilty to find the order by its documentNo or ID.
Before this change only was possible to find the order by the input orderid property
of the request json.
Now this behaviour still, but if the json has a documentNo property then the order ID
will be retrieved from the order with the provided documnet number. If the documentNo
is provided and there isn't any order with this documentNo then an exception is thrown.

--HG--
extra : rebase_source : d49ebcff1f10474cc78b17936224b7673cee946a