fixed BUG-43581: Create Lines From support for trigger exceptions
In case of errors during the process we rollback and close the current transaction. This is a general improvement that makes sense to the process.

In the JSONObject returned to the UI in case of exceptions, we try to get the message text to display to the end user. This is only possible because the previous locked transaction is closed, so DAL opens a new one to retrieve the error message.

Note that this code works for both db and java exceptions