Fixes ISSUE-43396: Multiselectors don't keep selection on multiple pages
In a multiple selectors, when a selected record was not in first page
and the selector was reopened, the record was not keep as selected. This
was becasue selectors datasource request assumes all selected records
were returned and selected records were not sorted on top of the rest.

This is fixed by using the same method P&E multiselection uses to:
  * Sort selected records on top of the rest
  * Check whether the full selection fits in the page size to increase
    it if neccessary to ensure the whole selection is returned in the
    request.