Fixes ISSUE-39212: Remove obsolete db.apply.modules & apply.on.update
Main build.xml file has some obsolete items which are never used and are
removed here.
Task db.apply.modules is never called anywhere in the build system.
Also it is not useful to keep for potential manual developer usage as
its purpose is just to call another ant target apply.module depending
on if the flag apply.on.update is set.
Removing that task also some small follow up cleanup on the condition
definition in + the properly definition.