Fixed ISSUE-43802. Make referencedinventory test auto-skip if AWO is installed.
Those test are not compatible with AWO as data they create does not comply with
some new constraint added by AWO.
So far that was required to be handled manually by adding config/disabled-tests
i.e. in every CI job with AWO installed to list the all individual tests here
to skip them.

This issue makes that behavior automatic by checking inside a central base-class
if AWO module is present and skipping the tests themselves (using similar logic
as is used by code for skipping config/disabled-tests from OBBaseTest).
Note that the repeated (per test) calls to isModulePresent are very fast as
that function caches the data inside so no need for caching isAwoInstalled inside
the test-class.