Fixes ISSUE-43963: Update param window multiselector param correctly on Done
When the form of a process definition is validated the array value of the multiselector parameter gets modified to string after its isString default validator is executed.
This happens, for example after pressing the "Done" button to execute the process.
This fix ensures that the value of the multiselector parameter is reset to the corresponding array of values that the multiselector has selected once the validation has been executed.