[indexeddb-hwm-endpoint] Additional adjustments
  - When converting Ticket state -> backbone force to have an id,
otherwise an empty backbone order is created instead of an order with
the information of the state Ticket.

  - If there is an error synchronizing the differents message in the
HardwareManagerEndpoint catch the message and log the error. This way we
are preventing to accumulate error messages in the database.

  - When reloading a backbone order we need to initialize properly the
calculatedInvoice property as a backbone order. Otherwise this property
is not initalized properly when converting state -> backbone which
causes error when trying to print a ticket with that property set.