[indexeddb-hwm-endpoint] Be consistent in ticket line printing
  It was confusing and inconsistent to make use of the
BackwardsCompatibility layer to parse backbone order lines and then not
use it for deserializing that information in the HardwareManagerEndpoint
when consuming the "printTicketLine" messages.

  Besides, no property mapping transformation was being done in that
case.

  Anyway, as we don't need to apply any transformation because we just
need to serialize/deserialize the backbone order line information itself
we are just parsing to JSON the backbone order lines to include them
into the messages without making use of the BackwardsCompatibility layer.