fixed FR-44164: make it possible to add info to process definition payload
Before it was possible for a process definintion to include additional
parameters calculated in a JavaScript function by using the view's
externalParams property.

This feature could not be used to add information to the payload based
on current paramteters as the way to inject JavaScript code is through a
client side validation which is executed after the payload is
calucalted.

Now additional paramters are added to the payload just before invoking
the server, in this way client side validation can be used also to add
information to the payload.