[indexeddb-hwm-endpoint] Allow to skip select printer popup
  When the terminal is configured to always show the select printer
window, there are now some cases where we want to prevent showing that
popup.

  This is the case of the "pay open tickets". That feature allows to pay
multiple tickets which eventually results in the printout of each of the
tickets. In this case we just one to show the popup just once, when the
first ticket is about to be printed.

  Added a new print argument "skipSelectPrinters" that allows to not
show the select printers popup even when the "selectprinteralways"
configuration is enabled.