Fixes ISSUE-44281: Use exists in hql filter clause of User window
HQL filter clause of User Window was previously using IN to filter from
BusinessPartner table, this resulted in a big hit in performance on
instances with a large number of BussinessPartners.

This has been fixed by using an exists and joining the table
BusinessPartner on the id being checked. It is similar to the SQL Filter
clause.