Fixes ISSUE-44525: wrong customer input is not set on slow connections
The problem was that on slow connections the cache of some combos was not the expected one.

On slow connection:

{_identifier: "Alimentos y Supermercados, S.A", id: "A6750F0D15334FB890C254369AC750A8", bpid: "A6750F0D15334FB890C254369AC750A8"}
1: "loading"
2: "loading"
3: "loading"...

On standard connection:

0: {category: "Customer - Tier 1", creditAvailable: 0, value: "ES-C1/0001", creditUsed: 0, name: "Alimentos y Supermercados, S.A", …}
1: {category: "Customer - Tier 3", creditAvailable: 0, value: "ES-C3/0001", creditUsed: 0, name: "Hoteles Buenas Noches, S.A.", …}
2: {category: "Customer - Tier 2", creditAvailable: 0, value: "ES-C2/0001", creditUsed: 0, name: "Restaurantes Luna Llena, S.A.", …}
3: "loading"
4: "loading"

Note that on standard connection the entries have a name property, while on slow connection they only contained an identifier, an id and bpid.

When the mapDisplayToValue(null) method was invoked on this combo, on slow connection it returned the first record, while on a standard connection
it returned null. This function was already being overwritten in ob-selector-item and returned null if the parameter was an empty string, now
it returns null if the parameter is null too